Fix sign/unsigned warnings in benchmain.cpp and gmmain.cpp
git-svn-id: http://skia.googlecode.com/svn/trunk@8079 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ gm/gmmain.cpp
@@ -1181,7 +1181,7 @@ int tool_main(int argc, char** argv) {
 
 #if SK_SUPPORT_GPU
     struct {
-        size_t  fBytes;
+        int     fBytes;
         int     fCount;
     } gpuCacheSize = { -1, -1 }; // -1s mean use the default
 #endif
@@ -1523,7 +1523,7 @@ int tool_main(int argc, char** argv) {
                     int count;
                     gr->getTextureCacheLimits(&count, &bytes);
                     if (-1 != gpuCacheSize.fBytes) {
-                        bytes = gpuCacheSize.fBytes;
+                        bytes = static_cast<size_t>(gpuCacheSize.fBytes);
                     }
                     if (-1 != gpuCacheSize.fCount) {
                         count = gpuCacheSize.fCount;
